requirements.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
flatbuffers==24.3.25
fonttools==4.51.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
idna==3.7
imageio==2.34.1
joblib==1.4.2
keras==3.3.3
keras-tuner==1.4.7
kiwisolver==1.4.5
kt-legacy==1.0.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
networkx==3.3
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
protobuf==4.25.3
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
rich==13.7.1
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
sequential==1.0.0
setuptools==69.5.1
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.5.10
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
